packages/api/pyproject.toml
[project]
name = "kynari-api"
version = "0.1.0"
description = "Kynari — Privacy-first emotion detection API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "psycopg[binary]>=3.2.0",
    "psycopg-pool>=3.2.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.28.0",
    "stripe>=11.0.0",
    "anthropic>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
